rework constants and types
