+++ libs/components/NetworkSelect.tsx
@@ -2,7 +2,7 @@ import { FC } from "react";
 import { css } from "@emotion/react";
 import { MenuItem, Select, SelectChangeEvent, Theme } from "@mui/material";
 import ExpandMore from "@mui/icons-material/ExpandMore";
-import { CENNZnetNetwork } from "@/types";
+import { CENNZnetNetwork } from "@/libs/types";
 
 const NIKAU: CENNZnetNetwork = "Nikau";
 const RATA: CENNZnetNetwork = "Rata";
