+++ libs/utils/index.ts
@@ -1,5 +1,6 @@
 export { EndowedAccounts } from "@/libs/utils/EndowedAccounts";
 export { errMsgFilter } from "@/libs/utils/errorHandling";
-export { supportedTokens } from "@/libs/utils/supportedTokens";
+export { default as fetchClaimStatus } from "@/libs/utils/fetchClaimStatus";
+export { default as setNewClaim } from "@/libs/utils/setNewClaim";
 export { default as fetchValidAccount } from "@/libs/utils/fetchValidAccount";
 export { default as supplyAccount } from "@/libs/utils/supplyAccount";
