--- libs/utils/supportedTokens.ts
@@ -1,14 +1,0 @@
-import { CENNZnetToken } from "@/types";
-
-export const supportedTokens: CENNZnetToken[] = [
-	{
-		symbol: "CENNZ",
-		assetId: 16000,
-		logo: "cennz.svg",
-	},
-	{
-		symbol: "CPAY",
-		assetId: 16001,
-		logo: "cpay.svg",
-	},
-];
