--- types/next-auth.d.ts
@@ -1,11 +1,0 @@
-import NextAuth, { DefaultSession } from "next-auth";
-
-declare module "next-auth" {
-	interface Session {
-		user: {
-			address: string;
-		} & DefaultSession["user"];
-		username?: string;
-		validAccount?: Boolean;
-	}
-}
